implemented FlameTransformation using uint instead of size_t

+++ src/color/InterpolatedPalette.cpp
@@ -6,13 +6,13 @@
 #include "InterpolatedPalette.h"
 
 InterpolatedPalette::InterpolatedPalette(std::vector<Color> palette) {
-    const size_t N = palette.size();
+    const unsigned int N = palette.size();
 
     if(N < 2) {
         throw std::invalid_argument("new palette must contain at least 2 colors");
     }
 
-    for(size_t i = 0; i < N; i++) {
+    for(unsigned int i = 0; i < N; i++) {
         this->palette.push_back(palette.at(i));
     }
 }
